[EventGrid] Split tests to enable min/max testing
This change splits our tests into two folders:

- `public` which only uses surface area exported by the package.
- `internal` which tests internal methods which we don't export from
   the top level package.

This allows us to test the public surface area independently from the
private one, which is helpful for classes of testing.

Fixes #13813
