+++ sdk/eventgrid/eventgrid/tests.yml
@@ -5,5 +5,3 @@ stages:
     parameters:
       PackageName: "@azure/eventgrid"
       ServiceDirectory: eventgrid
-      MatrixFilters:
-        - DependencyVersion=^$
